# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Это пример сервера с использованием Swagger в Gin.",
        "title": "swagger API",
        "contact": {},
        "version": "1.0"
    },
    "host": "157.90.227.125:8000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/admin/post": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post"
                ],
                "parameters": [
                    {
                        "description": "CreatePost data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.CreatePost"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-model_Post"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    }
                }
            }
        },
...
        "/api/v1/post": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post"
                ],
                "summary": "список группированных постов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "hashtag_id",
                        "name": "hashtag_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "collection_id",
                        "name": "collection_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "language",
                        "name": "language",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_PostResponseByGroup"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post"
                ],
                "parameters": [
                    {
                        "description": "UserLogin data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.UpdatePost"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-model_Post"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "post_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    }
                }
            }
        },
...
        "/api/v1/product": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "список товаров",
                "parameters": [
                    {
                        "type": "string",
                        "description": "product_tag_id",
                        "name": "product_tag_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-array_model_Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    }
                }
            }
        },
...
        "/api/v2/post": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "post"
                ],
                "summary": "список группированных постов",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "hashtag_id",
                        "name": "hashtag_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "collection_id",
                        "name": "collection_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "language",
                        "name": "language",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "size",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_PostResponseByGroup"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.Response-schema_Empty"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.CreatePost": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "company": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "image": {
                    "$ref": "#/definitions/admin.Image"
                },
                "language": {
                    "type": "string"
                },
                "logo": {
                    "$ref": "#/definitions/admin.Image"
                },
                "point": {
                    "type": "integer"
                },
                "quiz_time": {
                    "type": "integer"
                },
                "rating_status": {
                    "type": "string"
                },
                "read_time": {
                    "type": "integer"
                },
                "sapphire": {
                    "type": "integer"
                },
                "short_description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "admin.Image": {
            "type": "object",
            "properties": {
                "file": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                }
            }
        },
...
        "admin.UpdatePost": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "company": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "image": {
                    "$ref": "#/definitions/admin.Image"
                },
                "language": {
                    "type": "string"
                },
                "logo": {
                    "$ref": "#/definitions/admin.Image"
                },
                "point": {
                    "type": "integer"
                },
                "post_id": {
                    "type": "integer"
                },
                "quiz_time": {
                    "type": "integer"
                },
                "rating_status": {
                    "type": "string"
                },
                "read_time": {
                    "type": "integer"
                },
                "sapphire": {
                    "type": "integer"
                },
                "short_description": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "model.Collection": {
            "type": "object",
            "properties": {
                "collection_id": {
                    "type": "integer"
                },
                "image_path": {
                    "type": "string"
                },
                "image_path_kz": {
                    "type": "string"
                },
                "image_path_ru": {
                    "type": "string"
                },
                "is_recommendation": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "name_kz": {
                    "type": "string"
                },
                "name_ru": {
                    "type": "string"
                },
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Post"
                    }
                }
            }
        },
...
        "model.Hashtag": {
            "type": "object",
            "properties": {
                "hashtag_id": {
                    "type": "integer"
                },
                "image_path": {
                    "type": "string"
                },
                "is_visible": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "name_kz": {
                    "type": "string"
                },
                "name_ru": {
                    "type": "string"
                },
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Post"
                    }
                }
            }
        },
        "model.Image": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "image_id": {
                    "type": "integer"
                },
                "post_id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "model.Post": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "collections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Collection"
                    }
                },
                "company": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "hashtags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Hashtag"
                    }
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Image"
                    }
                },
                "language": {
                    "type": "string"
                },
                "point": {
                    "type": "integer"
                },
                "post_id": {
                    "type": "integer"
                },
                "post_order": {
                    "type": "integer"
                },
                "quiz_time": {
                    "type": "integer"
                },
                "rating_status": {
                    "type": "string"
                },
                "read_time": {
                    "type": "integer"
                },
                "sapphire": {
                    "type": "integer"
                },
                "short_description": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "model.Product": {
            "type": "object",
            "properties": {
                "airtable_product_id": {
                    "type": "string"
                },
                "contacts": {
                    "type": "string"
                },
                "contacts_en": {
                    "type": "string"
                },
                "contacts_kz": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "description_en": {
                    "type": "string"
                },
                "description_kz": {
                    "type": "string"
                },
                "discount": {
                    "type": "string"
                },
                "discount_en": {
                    "type": "string"
                },
                "discount_kz": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Image"
                    }
                },
                "offer": {
                    "type": "string"
                },
                "offer_en": {
                    "type": "string"
                },
                "offer_kz": {
                    "type": "string"
                },
                "point": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "product_tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ProductTag"
                    }
                },
                "product_type": {
                    "type": "string"
                },
                "sapphire": {
                    "type": "integer"
                },
                "sell_type": {
                    "type": "string"
                },
                "sku": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "title_en": {
                    "type": "string"
                },
                "title_kz": {
                    "type": "string"
                }
            }
        },
        "model.ProductTag": {
            "type": "object",
            "properties": {
                "image_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "name_kz": {
                    "type": "string"
                },
                "name_ru": {
                    "type": "string"
                },
                "product_tag_id": {
                    "type": "integer"
                }
            }
        },
...
        "schema.Empty": {
            "type": "object"
        },
...
        "schema.PostResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "collections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Collection"
                    }
                },
                "company": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "hashtags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Hashtag"
                    }
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Image"
                    }
                },
                "is_already_read": {
                    "type": "boolean"
                },
                "is_marked": {
                    "type": "boolean"
                },
                "language": {
                    "type": "string"
                },
                "mark_id": {
                    "type": "integer"
                },
                "point": {
                    "type": "integer"
                },
                "post_id": {
                    "type": "integer"
                },
                "post_order": {
                    "type": "integer"
                },
                "post_type": {
                    "type": "string"
                },
                "quiz_passed": {
                    "type": "boolean"
                },
                "quiz_time": {
                    "type": "integer"
                },
                "rating_status": {
                    "type": "string"
                },
                "read_time": {
                    "type": "integer"
                },
                "sapphire": {
                    "type": "integer"
                },
                "short_description": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "schema.PostResponseByGroup": {
            "type": "object",
            "properties": {
                "bestsellers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.PostResponse"
                    }
                },
                "continue_reading": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.PostResponse"
                    }
                },
                "partners": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.PostResponse"
                    }
                }
            }
        },
...
        "schema.Response-array_model_Product": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Product"
                    }
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "schema.Response-model_Post": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/model.Post"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "schema.Response-schema_Empty": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/schema.Empty"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "schema.Response-schema_PostResponseByGroup": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "$ref": "#/definitions/schema.PostResponseByGroup"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}